Refactor page cache with modular design and improved provenance tracking
