+++ src/praga_core/page_cache/__init__.py
@@ -1,0 +1,11 @@
+"""SQL-based page cache for storing and retrieving Page instances.
+
+This package provides a PageCache class that automatically creates SQL tables
+from Pydantic Page models and provides type-safe querying capabilities with
+provenance tracking support.
+"""
+
+from .core import PageCache
+from .exceptions import ProvenanceError
+
+__all__ = ["PageCache", "ProvenanceError"]
