+++ src/praga_core/page_cache/exceptions.py
@@ -1,0 +1,6 @@
+"""Exceptions for page cache operations."""
+
+
+class ProvenanceError(Exception):
+    """Exception raised for provenance tracking violations."""
+    pass
